+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- (Dis)advantage flags on saves against specific damage types.
+
+### Fixed
+
+- SW5E specific (dis)advantage flags should now work properly with midi-qol.
+
+### Removed
+
+- Character flag `sonicSensitivity`.
+
 ## [2.3.1.2.6.5] - 2023-11-15
 
 ### Added